pyproject.toml
[tool.poetry]
name = "entropix"
version = "0.1.0"
description = ""
authors = ['cole-tulip']
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tiktoken = "0.4.0"
pydantic = "^2.9.2"
blobfile = "^3.0.0"
ml-dtypes = "^0.5.0"
rich = "^13.8.1"
chex = "^0.1.87"
tyro = "^0.8.11"
jax = {extras = ["cuda12"], version = "^0.4.35"}
pynvml = "^11.0.0"
GPUtil = "^1.4.0"
psutil = "^5.9.5"
python-dotenv = "^1.0.0"
openai = "^1.55.1"
qdrant-client = "^1.12.1"
numpy = "^2.1.3"
tqdm = "^4.67.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"
transformers = "^4.45.1"
huggingface-hub = {extras = ["cli"], version = "^0.25.1"}


[tool.poetry.group.test.dependencies]
torch = "^2.4.1"
fairscale = "^0.4.13"

[tool.poetry.group.eval.dependencies]
lm-eval = "^0.4.4"
langdetect = "^1.0.9"
immutabledict = "^4.2.0"
anthropic = "^0.39.0"
requests = "^2.32.3"
pandas = "^2.2.3"
blobfile = "^3.0.0"

[tool.poetry.group.embeddings.dependencies]
transformers = "^4.45.1"
torch = {version = "^2.0.0", extras = ["cuda12"]}
qdrant-client = "^1.12.1"
python-dotenv = "^1.0.0"
tqdm = "^4.67.1"
einops = "^0.7.0"
openai = "^1.55.1"
backoff = "^2.2.1"

[tool.poetry.group.server.dependencies]
fastapi = "^0.115.3"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"
openai = "^1.52.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ["tests/**"]
lint.select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # Ruff
  "RUF",
]
lint.ignore = [
  "E501",   # ignore line-length (e.g. long docstring)
  "SIM117", # ignore combine with statements (it leads to harder to read code)
]
line-length = 88
indent-width = 2